+++ Project-Perfect-Citizen/Code/Game/createWindow.h
@@ -7,7 +7,7 @@
 
 namespace ppc {
 
-WindowInterface* spawnConsole(InputHandler& ih, NodeState& ns, int x, int y, int w, int h);
+WindowInterface* spawnConsole(InputHandler& ih, NodeState& ns, float x, float y, unsigned int w, unsigned int h);
 
 };
